finish prune-useless-value function

+++ src/Patient.java
@@ -4,7 +4,7 @@ This class represent a record in the data set
 public class Patient {
     private int id;
     private String gender;
-    private int age;
+    private double age;
     private boolean hypertension;
     private boolean heart_disease;
     private boolean ever_married;
@@ -18,7 +18,7 @@ public class Patient {
     public Patient(String[] metadata) {
         this.id = Integer.parseInt(metadata[0]);
         this.gender = metadata[1];
-        this.age = Integer.parseInt(metadata[2]);
+        this.age = Double.parseDouble(metadata[2]);
         this.hypertension = metadata[3].equals("1");
         this.heart_disease = metadata[4].equals("1");
         this.ever_married = metadata[5].equals("1");
@@ -41,7 +41,7 @@ public class Patient {
         return gender;
     }
 
-    public int getAge() {
+    public double getAge() {
         return age;
     }
